+++ Project.toml
@@ -4,7 +4,6 @@ authors = ["Leandro Martinez <lmartine@unicamp.br> and contributors"]
 version = "2.6.1-DEV"
 
 [deps]
-TestItems = "1c621080-faea-4a02-84b6-bbd5e436b8fe"
 
 [compat]
 Aqua = "0.8"
@@ -23,6 +22,7 @@ Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 TestItemRunner = "f8b46487-2199-4994-9208-9a1283c18c0a"
+TestItems = "1c621080-faea-4a02-84b6-bbd5e436b8fe"
 
 [targets]
-test = ["Test", "TestItemRunner", "OffsetArrays", "Aqua", "Documenter", "BenchmarkTools"]
+test = ["Test", "TestItemRunner", "TestItems", "OffsetArrays", "Aqua", "Documenter", "BenchmarkTools"]
